Refactor ticket detail page to Jira-style view
